refactor: split dir/builder.rs up
* split off dir_builder.rs
* split off PostOrderIterator
* split bufferedtreebuilder off
