+++ unixfs/src/dir/builder/buffered.rs
@@ -1,0 +1,388 @@
+use super::{DirBuilder, Entry, Leaf, PostOrderIterator, TreeBuildingFailed, TreeOptions, Visited};
+use crate::Metadata;
+use cid::Cid;
+use std::collections::hash_map::Entry::*;
+
+/// UnixFs directory tree builder, which buffers entries until `build()` is called.
+#[derive(Debug)]
+pub struct BufferingTreeBuilder {
+    /// At the root there can be only one element, unless an option was given to create a new
+    /// directory surrounding the root elements.
+    root_builder: DirBuilder,
+    longest_path: usize,
+    // used to generate each node an unique id which is used when doing the post order traversal to
+    // recover all childrens rendered Cids
+    counter: u64,
+    opts: TreeOptions,
+}
+
+impl Default for BufferingTreeBuilder {
+    fn default() -> Self {
+        Self::new(TreeOptions::default())
+    }
+}
+
+impl BufferingTreeBuilder {
+    /// Construct a new tree builder with the given configuration.
+    pub fn new(opts: TreeOptions) -> Self {
+        BufferingTreeBuilder {
+            root_builder: DirBuilder::root(0),
+            longest_path: 0,
+            counter: 1,
+            opts,
+        }
+    }
+
+    /// Records the give path to be a link to the following cid.
+    ///
+    /// FIXME: this should be renamed as "put_leaf" or "put_opaque_leaf".
+    pub fn put_file(
+        &mut self,
+        full_path: &str,
+        target: Cid,
+        total_size: u64,
+    ) -> Result<(), TreeBuildingFailed> {
+        // inserted at the depth
+        let leaf = Leaf {
+            link: target,
+            total_size,
+        };
+
+        self.modify_with(full_path, |parent, basename, _| {
+            parent
+                .put_leaf(basename, leaf)
+                .map_err(|_| TreeBuildingFailed::DuplicatePath(full_path.to_string()))
+        })
+    }
+
+    /// Directories get "put" implicitly through the put files, and directories need to be adjusted
+    /// only when wanting them to have metadata.
+    pub fn set_metadata(
+        &mut self,
+        full_path: &str,
+        metadata: Metadata,
+    ) -> Result<(), TreeBuildingFailed> {
+        // create all paths along the way
+        //
+        // set if not set, error otherwise? FIXME: doesn't error atm
+        self.modify_with(full_path, |parent, basename, id| {
+            parent
+                .add_or_get_node(basename, id)
+                .map_err(|_| TreeBuildingFailed::LeafAsDirectory(full_path.to_string()))?
+                .set_metadata(metadata);
+            Ok(())
+        })
+    }
+
+    fn modify_with<F>(&mut self, full_path: &str, f: F) -> Result<(), TreeBuildingFailed>
+    where
+        F: FnOnce(&mut DirBuilder, String, &mut Option<u64>) -> Result<(), TreeBuildingFailed>,
+    {
+        // create all paths along the way
+        //
+        // assuming it's ok to split '/' since that cannot be escaped in linux at least
+
+        self.longest_path = full_path.len().max(self.longest_path);
+        let mut remaining = full_path.split('/').enumerate().peekable();
+        let mut dir_builder = &mut self.root_builder;
+
+        // needed to avoid borrowing into the DirBuilder::new calling closure
+        let counter = &mut self.counter;
+
+        while let Some((depth, next)) = remaining.next() {
+            let last = remaining.peek().is_none();
+
+            match (depth, next, last) {
+                // this might need to be accepted in case there is just a single file
+                (0, "", true) => { /* accepted */ }
+                (0, "", false) => {
+                    return Err(TreeBuildingFailed::RootedPath(full_path.to_string()))
+                }
+                (_, "", false) => {
+                    return Err(TreeBuildingFailed::RepeatSlashesInPath(
+                        full_path.to_string(),
+                    ))
+                }
+                (_, "", true) => todo!("path ends in slash"),
+                _ => {}
+            }
+
+            // our first level can be full given the options
+            let full = depth == 0 && !self.opts.wrap_in_directory && dir_builder.is_empty();
+
+            if last {
+                let mut next_id = Some(*counter);
+
+                let ret = if full {
+                    Err(TreeBuildingFailed::TooManyRootLevelEntries)
+                } else {
+                    f(dir_builder, next.to_string(), &mut next_id)
+                };
+
+                if next_id.is_none() {
+                    *counter += 1;
+                }
+
+                if ret.is_err() {
+                    // FIXME: there might be a case where we have now stale nodes in our tree but
+                    // cannot figure out an example for that.
+                }
+
+                return ret;
+            }
+
+            let parent_id = dir_builder.id;
+
+            dir_builder = match (full, dir_builder.nodes.entry(next.to_string())) {
+                (_, Occupied(oe)) => oe
+                    .into_mut()
+                    .as_dir_builder()
+                    .map_err(|_| TreeBuildingFailed::LeafAsDirectory(full_path.to_string()))?,
+                (false, Vacant(ve)) => {
+                    let next_id = *counter;
+                    *counter += 1;
+                    ve.insert(Entry::Directory(DirBuilder::new(parent_id, next_id)))
+                        .as_dir_builder()
+                        .expect("safe: we just inserted a DirBuilder")
+                }
+                (true, Vacant(_)) => return Err(TreeBuildingFailed::TooManyRootLevelEntries),
+            };
+        }
+
+        // as the str::split will always return a single element this should not ever be hit
+        unreachable!(
+            "walked the full_path but failed to add anything: {:?}",
+            full_path
+        );
+    }
+
+    /// Called to build the tree. The built tree will have the added files and their implied
+    /// directory structure, along with the any directory entries which were created using
+    /// `set_metadata`. To build the whole hierarchy, one must iterate the returned iterator to
+    /// completion while storing the created blocks.
+    ///
+    /// Returned `PostOrderIterator` will use the given `full_path` and `block_buffer` to store
+    /// it's data during the walk. `PostOrderIterator` implements `Iterator` while also allowing
+    /// borrowed access via `next_borrowed`.
+    pub fn build<'a>(
+        self,
+        full_path: &'a mut String,
+        block_buffer: &'a mut Vec<u8>,
+    ) -> PostOrderIterator<'a> {
+        PostOrderIterator::new(
+            Visited::Descent {
+                node: self.root_builder,
+                name: None,
+                depth: 0,
+            },
+            full_path,
+            block_buffer,
+            self.opts,
+        )
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{BufferingTreeBuilder, Metadata, TreeBuildingFailed, TreeOptions};
+    use cid::Cid;
+    use std::convert::TryFrom;
+
+    #[test]
+    fn some_directories() {
+        let mut builder = BufferingTreeBuilder::default();
+
+        // foobar\n
+        let five_block_foobar =
+            Cid::try_from("QmRJHYTNvC3hmd9gJQARxLR1QMEincccBV53bBw524yyq6").unwrap();
+
+        builder
+            .put_file("a/b/c/d/e/f/g.txt", five_block_foobar.clone(), 221)
+            .unwrap();
+        builder
+            .put_file("a/b/c/d/e/h.txt", five_block_foobar.clone(), 221)
+            .unwrap();
+        builder
+            .put_file("a/b/c/d/e/i.txt", five_block_foobar.clone(), 221)
+            .unwrap();
+
+        let mut full_path = String::new();
+        let mut buffer = Vec::new();
+
+        let iter = builder.build(&mut full_path, &mut buffer);
+        let mut actual = iter
+            .map(|res| res.map(|(p, cid, buf)| (p, cid, buf)))
+            .collect::<Result<Vec<_>, _>>()
+            .unwrap();
+
+        let mut expected = vec![
+            (
+                "a/b/c/d/e/f",
+                "Qmbgf44ztW9wLcGNRNYGinGQB6SQDQtbHVbkM5MrWms698",
+            ),
+            (
+                "a/b/c/d/e",
+                "Qma1hCr3CuPRAq2Gw4DCNMqsi42Bjs4Bt1MGSS57kNh144",
+            ),
+            ("a/b/c/d", "QmUqaYatcJqiSFdykHXGh4Nog1eMSfDJBeYzcG67KV5Ri4"),
+            ("a/b/c", "QmYwaNBaGpDCNN9XpHmjxVPHmEXZMw9KDY3uikE2UU5fVB"),
+            ("a/b", "QmeAzCPig4o4gBLh2LvP96Sr8MUBrsu2Scw9MTq1EvTDhY"),
+            ("a", "QmSTUFaPwJW8xD4KNRLLQRqVTYtYC29xuhYTJoYPWdzvKp"),
+        ];
+
+        // hopefully this way the errors will be easier to hunt down
+
+        actual.reverse();
+        expected.reverse();
+
+        while let Some(actual) = actual.pop() {
+            let expected = expected.pop().expect("size mismatch");
+            assert_eq!(actual.0, expected.0);
+            assert_eq!(
+                actual.1.to_string(),
+                expected.1,
+                "{:?}: {:?}",
+                actual.0,
+                Hex(&actual.2)
+            );
+        }
+    }
+
+    struct Hex<'a>(&'a [u8]);
+    use std::fmt;
+
+    impl<'a> fmt::Debug for Hex<'a> {
+        fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
+            for b in self.0 {
+                write!(fmt, "{:02x}", b)?;
+            }
+            Ok(())
+        }
+    }
+
+    #[test]
+    fn empty_path() {
+        let mut builder = BufferingTreeBuilder::default();
+        builder.put_file("", some_cid(0), 1).unwrap();
+
+        let mut full_path = String::new();
+        let mut buffer = Vec::new();
+
+        let iter = builder.build(&mut full_path, &mut buffer);
+        let actual = iter
+            .map(|res| res.map(|(p, _, _)| p))
+            .collect::<Result<Vec<_>, _>>()
+            .unwrap();
+
+        assert!(
+            actual.is_empty(),
+            "wrapping in directory was not asked, single element"
+        );
+    }
+
+    #[test]
+    #[should_panic]
+    fn rooted_path() {
+        let mut builder = BufferingTreeBuilder::default();
+        builder.put_file("/a", some_cid(0), 1).unwrap();
+    }
+
+    #[test]
+    #[should_panic]
+    fn successive_slashes() {
+        let mut builder = BufferingTreeBuilder::default();
+        builder.put_file("a//b", some_cid(0), 1).unwrap();
+    }
+
+    #[test]
+    fn multiple_roots() {
+        // foobar\n
+        let five_block_foobar =
+            Cid::try_from("QmRJHYTNvC3hmd9gJQARxLR1QMEincccBV53bBw524yyq6").unwrap();
+
+        let opts = TreeOptions::default().with_wrap_in_directory();
+        let mut builder = BufferingTreeBuilder::new(opts);
+        builder
+            .put_file("a", five_block_foobar.clone(), 221)
+            .unwrap();
+        builder.put_file("b", five_block_foobar, 221).unwrap();
+
+        let mut full_path = String::new();
+        let mut buffer = Vec::new();
+
+        let iter = builder.build(&mut full_path, &mut buffer);
+        let actual = iter
+            .map(|res| res.map(|(p, _, _)| p))
+            .collect::<Result<Vec<_>, _>>()
+            .unwrap();
+
+        assert_eq!(actual, &["QmdbWuhpVCX9weVMMqvVTMeGwKMqCNJDbx7ZK1zG36sea7"]);
+    }
+
+    #[test]
+    #[should_panic]
+    fn denied_multiple_root_dirs() {
+        let mut builder = BufferingTreeBuilder::default();
+        builder.put_file("a/c.txt", some_cid(0), 1).unwrap();
+        builder.put_file("b/d.txt", some_cid(1), 1).unwrap();
+    }
+
+    #[test]
+    #[should_panic]
+    fn denied_multiple_root_files() {
+        let mut builder = BufferingTreeBuilder::default();
+        builder.put_file("a.txt", some_cid(0), 1).unwrap();
+        builder.put_file("b.txt", some_cid(1), 1).unwrap();
+    }
+
+    #[test]
+    #[should_panic]
+    fn using_leaf_as_node() {
+        let mut builder = BufferingTreeBuilder::default();
+        builder.put_file("a.txt", some_cid(0), 1).unwrap();
+        builder.put_file("a.txt/b.txt", some_cid(1), 1).unwrap();
+    }
+
+    #[test]
+    fn set_metadata_before_files() {
+        let mut builder = BufferingTreeBuilder::default();
+        builder
+            .set_metadata("a/b/c/d", Metadata::default())
+            .unwrap();
+        builder.put_file("a/b/c/d/e.txt", some_cid(1), 1).unwrap();
+        builder.put_file("a/b/c/d/f.txt", some_cid(2), 1).unwrap();
+
+        let mut full_path = String::new();
+        let mut buffer = Vec::new();
+
+        let iter = builder.build(&mut full_path, &mut buffer);
+        let actual = iter
+            .map(|res| res.map(|(p, _, _)| p))
+            .collect::<Result<Vec<_>, _>>()
+            .unwrap();
+
+        assert_eq!(actual, &["a/b/c/d", "a/b/c", "a/b", "a",])
+    }
+
+    #[test]
+    fn set_metadata_on_file() {
+        let mut builder = BufferingTreeBuilder::default();
+        builder.put_file("a/a.txt", some_cid(0), 1).unwrap();
+        let err = builder
+            .set_metadata("a/a.txt", Metadata::default())
+            .unwrap_err();
+
+        assert!(
+            matches!(err, TreeBuildingFailed::LeafAsDirectory(_)),
+            "{:?}",
+            err
+        );
+    }
+
+    /// Returns a quick and dirty sha2-256 of the given number as a Cidv0
+    fn some_cid(number: usize) -> Cid {
+        use multihash::Sha2_256;
+        let mh = Sha2_256::digest(&number.to_le_bytes());
+        Cid::new_v0(mh).unwrap()
+    }
+}
